+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/INodeDirectory.java
@@ -51,8 +51,8 @@ public class INodeDirectory extends INode {
   protected static final int DEFAULT_FILES_PER_DIRECTORY = 5;
   public final static String ROOT_NAME = "";
   //START_HOP_CODE
-  public final static long ROOT_ID =  1L;
-  public final static long ROOT_PARENT_ID = NON_EXISTING_ID;
+  public final static int ROOT_ID =  1;
+  public final static int ROOT_PARENT_ID = NON_EXISTING_ID;
   //END_HOP_CODE
   
   //private List<INode> children;       // no need for a list here. get it from the DB
@@ -293,7 +293,7 @@ public class INodeDirectory extends INode {
     }
 
     if (!node.exists()) {
-      long inodeID = HopINodeIdGen.getUniqueINodeID();
+      Integer inodeID = HopINodeIdGen.getUniqueINodeID();
       node.setIdNoPersistance(inodeID);
       node.setParentNoPersistance(this);
       EntityManager.add(node);
